Cleaned up code for PR
