+++ package.json
@@ -3,9 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@nivo/line": "^0.33.0",
+    "dygraphs": "^2.1.0",
     "lodash": "^4.17.5",
-    "lodash-es": "^4.17.7",
     "material-ui": "^0.20.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
