--- src/Data/nivoData.js
@@ -1,258 +1,0 @@
-export default
-[
-  {
-    "id": "whisky",
-    "color": "hsl(127, 70%, 50%)",
-    "data": [
-      {
-        "color": "hsl(79, 70%, 50%)",
-        "x": "CW",
-        "y": 58
-      },
-      {
-        "color": "hsl(209, 70%, 50%)",
-        "x": "CM",
-        "y": 0
-      },
-      {
-        "color": "hsl(15, 70%, 50%)",
-        "x": "NL",
-        "y": 46
-      },
-      {
-        "color": "hsl(274, 70%, 50%)",
-        "x": "MG",
-        "y": 34
-      },
-      {
-        "color": "hsl(33, 70%, 50%)",
-        "x": "RW",
-        "y": 58
-      },
-      {
-        "color": "hsl(302, 70%, 50%)",
-        "x": "ML",
-        "y": 58
-      },
-      {
-        "color": "hsl(281, 70%, 50%)",
-        "x": "BH",
-        "y": 58
-      },
-      {
-        "color": "hsl(291, 70%, 50%)",
-        "x": "GP",
-        "y": 6
-      },
-      {
-        "color": "hsl(315, 70%, 50%)",
-        "x": "BE",
-        "y": 9
-      }
-    ]
-  },
-  {
-    "id": "rhum",
-    "color": "hsl(155, 70%, 50%)",
-    "data": [
-      {
-        "color": "hsl(230, 70%, 50%)",
-        "x": "CW",
-        "y": 38
-      },
-      {
-        "color": "hsl(6, 70%, 50%)",
-        "x": "CM",
-        "y": 42
-      },
-      {
-        "color": "hsl(315, 70%, 50%)",
-        "x": "NL",
-        "y": 5
-      },
-      {
-        "color": "hsl(340, 70%, 50%)",
-        "x": "MG",
-        "y": 29
-      },
-      {
-        "color": "hsl(164, 70%, 50%)",
-        "x": "RW",
-        "y": 54
-      },
-      {
-        "color": "hsl(141, 70%, 50%)",
-        "x": "ML",
-        "y": 25
-      },
-      {
-        "color": "hsl(44, 70%, 50%)",
-        "x": "BH",
-        "y": 52
-      },
-      {
-        "color": "hsl(179, 70%, 50%)",
-        "x": "GP",
-        "y": 40
-      },
-      {
-        "color": "hsl(75, 70%, 50%)",
-        "x": "BE",
-        "y": 33
-      }
-    ]
-  },
-  {
-    "id": "gin",
-    "color": "hsl(39, 70%, 50%)",
-    "data": [
-      {
-        "color": "hsl(88, 70%, 50%)",
-        "x": "CM",
-        "y": 53
-      },
-      {
-        "color": "hsl(141, 70%, 50%)",
-        "x": "CW",
-        "y": 58
-      },
-      {
-        "color": "hsl(77, 70%, 50%)",
-        "x": "NL",
-        "y": 28
-      },
-      {
-        "color": "hsl(301, 70%, 50%)",
-        "x": "MG",
-        "y": 54
-      },
-      {
-        "color": "hsl(96, 70%, 50%)",
-        "x": "RW",
-        "y": 2
-      },
-      {
-        "color": "hsl(142, 70%, 50%)",
-        "x": "ML",
-        "y": 43
-      },
-      {
-        "color": "hsl(141, 70%, 50%)",
-        "x": "BH",
-        "y": 18
-      },
-      {
-        "color": "hsl(332, 70%, 50%)",
-        "x": "GP",
-        "y": 48
-      },
-      {
-        "color": "hsl(286, 70%, 50%)",
-        "x": "BE",
-        "y": 10
-      }
-    ]
-  },
-  {
-    "id": "vodka",
-    "color": "hsl(141, 70%, 50%)",
-    "data": [
-      {
-        "color": "hsl(230, 70%, 50%)",
-        "x": "CW",
-        "y": 20
-      },
-      {
-        "color": "hsl(85, 70%, 50%)",
-        "x": "CM",
-        "y": 37
-      },
-      {
-        "color": "hsl(233, 70%, 50%)",
-        "x": "NL",
-        "y": 56
-      },
-      {
-        "color": "hsl(227, 70%, 50%)",
-        "x": "MG",
-        "y": 16
-      },
-      {
-        "color": "hsl(158, 70%, 50%)",
-        "x": "RW",
-        "y": 31
-      },
-      {
-        "color": "hsl(357, 70%, 50%)",
-        "x": "ML",
-        "y": 4
-      },
-      {
-        "color": "hsl(151, 70%, 50%)",
-        "x": "BH",
-        "y": 38
-      },
-      {
-        "color": "hsl(298, 70%, 50%)",
-        "x": "GP",
-        "y": 12
-      },
-      {
-        "color": "hsl(133, 70%, 50%)",
-        "x": "BE",
-        "y": 30
-      }
-    ]
-  },
-  {
-    "id": "cognac",
-    "color": "hsl(305, 70%, 50%)",
-    "data": [
-      {
-        "color": "hsl(72, 70%, 50%)",
-        "x": "CW",
-        "y": 13
-      },
-      {
-        "color": "hsl(18, 70%, 50%)",
-        "x": "CM",
-        "y": 32
-      },
-      {
-        "color": "hsl(233, 70%, 50%)",
-        "x": "NL",
-        "y": 39
-      },
-      {
-        "color": "hsl(239, 70%, 50%)",
-        "x": "MG",
-        "y": 49
-      },
-      {
-        "color": "hsl(123, 70%, 50%)",
-        "x": "RW",
-        "y": 7
-      },
-      {
-        "color": "hsl(313, 70%, 50%)",
-        "x": "ML",
-        "y": 38
-      },
-      {
-        "color": "hsl(197, 70%, 50%)",
-        "x": "BH",
-        "y": 39
-      },
-      {
-        "color": "hsl(154, 70%, 50%)",
-        "x": "GP",
-        "y": 34
-      },
-      {
-        "color": "hsl(171, 70%, 50%)",
-        "x": "BE",
-        "y": 41
-      },
-    ]
-  }
-]
